VERSION parser into separate module
